[Modify] 작성아이콘 컴포넌트인 WriteMemoIconButton을 RegistrationIcon으로 이름 변경
