+++ src/components/atoms/RegistrationIcon/RegistrationIcon.stories.js
@@ -1,0 +1,15 @@
+import RegistrationIcon from './RegistrationIcon';
+
+const metaConfig = {
+  title: 'components/RegistrationIcon',
+  component: RegistrationIcon,
+  tags: ['autodocs'],
+  args: {},
+};
+
+export default metaConfig;
+
+export const Base = {
+  args: {},
+};
+Base.storyName = 'RegistrationIcon';
