--- src/components/atoms/WriteMemoIconButton/WriteMemoIconButton.stories.js
@@ -1,15 +1,0 @@
-import WriteMemoIconButton from './WriteMemoIconButton';
-
-const metaConfig = {
-  title: 'components/WriteMemoIconButton',
-  component: WriteMemoIconButton,
-  tags: ['autodocs'],
-  args: {},
-};
-
-export default metaConfig;
-
-export const Base = {
-  args: {},
-};
-Base.storyName = '메모 작성';
